Fixed a bug where the clamp_to_limits function wasn't 'even', I should add a test for it

+++ src/gcode-machine-control.cc
@@ -631,7 +631,7 @@ static float clamp_to_limits(const float value,
   float ratio, max_offset = 1, offset;
   for (int i = 0; i < GCODE_NUM_AXES; ++i) {
     // Keep ratio, increase readability
-    ratio = fabs(axis_steps[i]) / axis_steps[defining_axis];
+    ratio = fabs((float) axis_steps[i] / axis_steps[defining_axis]);
     offset = ratio > 0 ? max_axis_value[i] / (value * ratio) : 1;
     if (offset < max_offset) max_offset = offset;
   }
